ML/pyproject.toml
[project]
name = "ml-server"
version = "0.1.0"
description = "Servidor gRPC con BETO y MLflow."
requires-python = ">=3.9"

dependencies = [
    "grpcio>=1.65.0",
    "grpcio-tools>=1.65.0",
    "transformers>=4.44.0",
    "torch>=2.3.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
index-url = "https://pypi.org/simple"
